Update draft-ietf-core-object-security.md
This addresses and closes #2 (Use of different identifiers. This aligns with EDHOC)
This addresses and closes #12 (Remove multicast)
This closes #15 (definition and use of sid)
This partly addresses #62 (Use of different identifiers remove the need to use random cids with EDHOC and significantly reduces 
per packet overhead).
This partly addresses #70 (Use of different identifiers remove the need to use random cids with EDHOC and reduces 
memory overhead).