Replace Sender IV and Recipient IV with a common Context IV
After reviewing the newly added section on "Nonce generation examples", reviewing the multicast draft, and thinking about ways to reduce the memory usage of OSCOAP I have come to the conclusion that I think the Sender IV and Recipient IV should be replaced with a common Context IV (in the Common Context) (I.e. a Fixed-Common field according to RFC5116). As each endpoint uses a unique key for encryption, a Fixed-Distinct field (RFC5116) is not necessary. 

+ This significantly reduces memory usage or processing (depending on the implementation) especially in the case of multicast or an endpoint with many security contexts. 

+ This simplifies the specification

+ This simplifies the implementation

+ This makes it easier to analyse the security.

The Context IV (Fixed field according to RFC5116) are used to protect against pre-computation and time-memory tradeoff attacks. A Common IV in the common context would fulfil this purpose.