Removed alg from AAD
The alg is already included in the key derivation, another algorithm would generate a different key and validation would fail.

If there is a reason to include this, then the signature algorithm should be included in group OSCORE, which it is not. Need to be neither or both I think.