IESG review AR: remove Figure 5 v2
Suggested by IESG to remove Figure 5 as is does not provide any information. I think it makes more sense to add payload to the header fields table rather than the options table. The figures are floats and does not belong to the section they happen to be placed. But I think that the IESG comments could be applied to the whole section. Suggest merging the sections instead.