Removed note about target resource
The current text was recently clarified per comment from Jim and needed even further clarifications since the text related to how the request was handled in an intermediary device which was not described. Being only a note it did not really merit such a long explanation and was instead removed.